Improve UX of resending email confirmation
**Why**: In the scenario where a user visited `/sign_up/email/confirm`
with an invalid token (missing, expired, not matching any user), after
submitting their email, they were redirected to the sign in page with
a flash message. This experience is not consistent with the new views
we've added that make it easy for a user to have the email sent again
or to enter a different email address.

**How**:
- Instead of redirecting to the root URL, redirect to
`/sign_up/verify_email`, which is an existing view that can be reused
for this scenario.
- Refactor `EmailResendController` to make it consistent with other
controllers that use a Form Object for submitting forms and capturing
analytics. This allows us to remove the `ValidEmailParameter` concern.