Allow locale switching via URL param
**Why**: To make it easier to switch between English and Spanish, and
to see what has and has not yet been translated.

**How**: Add a before action in ApplicationController to set the locale
based on the URL param if it is present, or fall back to the default.

Note that eventually, we might implement something more user friendly,
such as fetching the locale from the domain name, such as login.gov.es
or es.login.gov, or perhaps in the path, like login.gov/es/profile.