Don't load user_flows rake task in production (#975)
**Why**: When the `rake` command is invoked, it parses all rake tasks,
even if they're not being called directly. The user_flows rake task
requires RSpec, and is not meant to be run in production. By ignoring
production, we suppress error messages due to RSpec not being present,
which aren't relevant in production.

In development, we don't want to rescue the LoadError. We want to make
sure the rake task aborts because if RSpec is not present, it's a
problem.