Encrypt otp_secret_key
**Why**: The `otp_secret_key` is what is used to generate the codes for
authenticator apps. If someone got a hold of our DB, they would be able
to get the 6-digit TOTP code for any user who had enabled an
authenticator app.

**How**:
- Use our current encryption scheme to populate the
`encrypted_otp_secret_key` column and override the reader and writer
methods for `otp_secret_key` so that the value is read from the
encrypted column
- Change the `encrypted_otp_secret_key` column type from string to text
because the encrypted value is more than 255 characters
- Remove unused `encrypted_otp_secret_key_*` columns. They came from the
`two_factor_authentication` gem, but we're not using the gem's
encryption.

Once these migrations and code have been pushed to all servers,
we'll remove the `otp_secret_key` column.