Notify user via email of phone number change (#1148)
**Why**: Before, we were assuming that all 2FA phones would be
SMS-enabled, which is why we notified users of phone changes by
sending them an SMS to the old phone number. Now that we accept
landlines and other types of phones that can't receive SMS, we need
a more robust way to notify users of the change.

Email is the simplest solution, which is what is implemented here.