Remove unnecessary DB calls (#1151)
**Why**: To make the app run faster.

**How**: Devise automatically includes `before_validation` callbacks
that downcase and strip whitespace from attributes defined in the
`config.case_insensitive_keys` and `config.strip_whitespace_keys`
arrays. The `email` attribute is included by default. This means that
every time the User model is saved, even if the email wasn't updated,
a DB call will be made to downcase and strip whitespace from the email.

This is unnecessary and can be avoided by making those config arrays
empty. Note that we already downcase and strip emails, and only when
necessary (i.e. when the email attribute is being created or updated,
and when a user is entering an email address in a form). This is the
proper way to handle this formatting, as opposed to via a model callback
that performs this action regardless of whether or not it is needed.

I also overrode the `downcase_keys` and `strip_whitespace` methods to
do nothing.

I noticed these unnecessary DB calls by running the app with
`rack_mini_profiler` set to `on`.