Rescue AWS error when KMS on and password wrong (#1248)
**Why**: If a user signs in with a wrong password, the ciphertext blob
passed to the AWS KMS service will not match the blob that corresponds
to the correct password. AWS raises an error in this case, which causes
the user to see the 500 error page.

**How**: Rescue the AWS error and raise Pii::EncryptionError so that
`valid_password?` returns `false` and the user sees the regular
"Invalid email or password" flash message on the sign in page.

Note that the test is somewhat brittle because it assumes that AWS
will always handle this scenario by raising this exact error. If this
changes in the future, our test will still pass. We would either need
to keep up with all the changes to the API, or set up a KMS test
environment we can use to talk to the real API.