Allow signing in to SP after session timeout
**Why**: We rely on the original SP request URL to be available in the
session in order to complete the authentication flow and redirect the
user back to the SP. If a user comes from an SP, but sits on the sign
in page long enough for the session to time out, the request URL will
be deleted from the session, and after they sign back in, they will end
up on the profile page instead of going back to the SP.

**How**:
- Take advantage of the `request_id` param to restore the SP info
in the session after the user signs in, the same way we do when a user
confirms their email in a different browser
- Create a `signed_in_path` method that only redirects to the profile
path if the user is fully authenticated. Otherwise, it redirects to
user_two_factor_authentication_path. Use this new method in
`after_sign_in_path_for`. Why? Because any controller that includes a
`before_action` that gets executed will set the stored location for the
user to that controller. This means that if we redirect to the
profile_path while there is no stored location (which will be the case
if the session times out) and the user isn't yet fully authenticated, it
will call the `confirm_two_factor_authenticated` `before_action`, and
the stored location will now be the profile page. So, after the user
enters their OTP, they will be sent to the profile page instead of back
to SamlIdpController#auth to complete the auth flow back to the SP.