Fixes bug with `this` ref in focustrap deactivate
**Why**: I thought the `onDeactivate` callback that the focusTrap
library executes when a trap is deativated returned that trap.
It actually returns undefined, which means that `this` pointed to a
different object. The expected method wasn't available on that object,
so an error was thrown.

**How**: I removed our callback wrapper entirely, and use an array
instead of a single reference to determine which trap object should be
considered the last active for purposes of restoring focus.