Refactor personal key JS specs (#1282)
**Why**:
- The JS-enabled specs were being called in two places because
they were included in a shared example. There's no benefit to running
these slow tests twice.
- The session timeout modal spec was failing intermittently due to being
nested inside a context that included a before block that already signed
in the user, and the test itself was trying to sign in the user as well.

**How**:
- Move the JS specs from `shared_examples_for_recovery_codes.rb` to
`users/recovery_code_spec.rb`.
- Merge all the specs into one big spec to speed up the test, and define
methods that describe the actions and expectations for readability.