Check for SP logo file existence before rendering
**WHY**: We are going to start allowing agencies to tell us the name of
their logo. The process will still include adding the logo to the IDP
codebase separately. In the case that an agency enters in the wrong name
or sets up an SP logo before the image is deployed, we want to make sure
we don't return a broken image. This logic checks for whether the file
is there before rendering the image.

See
http://stackoverflow.com/questions/6782978/rails-3-1-determine-if-asset-exists/35319530#35319530
as ref for the "see if this asset exists" code.