Do not show DAP on "enter email" page
**Why**: Before, we were excluding from POST controller action, which
actually does not render a view. Instead, it redirects to various views
depending on whether the confirmation token is valid or not. If it is
invalid, we don't care about whether we show DAP or not because the
session does not have trusted info. If the user is already confirmed,
they will be logged in so DAP will not be included. So what we really
care about is just the case when a user has a valid confirmation token.
In that case, the user is redirected to `sign_up_enter_password_url`
(see `process_valid_confirmation_token` method). Now, we are excluding
DAP from that page.

QA'd locally to make sure.