Wait until password entry to set SP session data
**Why**: We've come across a reproducible scenario in Chrome where if
the SameSite cookie setting is `Strict`, clicking on the confirmation
email from a Gmail web page resets the IdP session after it gets set in
`SignUp::EmailConfirmationsController`. This means that after the user
finishes completing the account creation, they won't be redirected back
to the SP.

By making this change, it will hopefully allow us to revert back to the
more secure `Strict` setting because we had to change it to `Lax` to
get around this Chrome issue.