Replace Marshal with JSON serializer
**Why**: For security reasons. From Ruby's docs:

> Ruby's Marshal module provides methods for serializing and
deserializing Ruby object trees to and from a binary data format.

> Never use Marshal.load to deserialize untrusted or user supplied data.
Because Marshal can deserialize to almost any Ruby object and has full
control over instance variables, it is possible to craft a malicious
payload that executes code shortly after deserialization.

> If you need to deserialize untrusted data, you should use JSON as it
is only capable of returning 'primitive' types such as strings, arrays,
hashes, numbers and nil. If you need to deserialize other classes, you
should handle this manually. Never deserialize to a user specified
class.

Rails 4.1 switched the default cookie serializer from Marshal to JSON
for these reasons.

**How**:
- Since JSON converts Hash keys to Strings, and in order to minimize
the changes in this PR, we call `.with_indifferent_access` on the
deserialized Hash to be able to refer to keys via symbols, such as
`session[:sp][:issuer]`.
- Ensure only basic types are stored in the session. Otherwise, they
won't be properly deserialized. For example, before, we were storing
instances of Classes (such as `Proofer::Resolution`) in the session.
Now, we are only storing Strings, Integers, Arrays, and Booleans.
- Add some logic for backwards compatibility in the scenario that
this code gets deployed at the same time as people are using the site.
To prevent sessions that were active before the deploy from being
terminated after the deploy, we check to see if the session was based
on Marshal, and if so, deserialize it, then serialize it with JSON.