Preserve request_id param across multiple visits
**Why**: Before, we were passing in the request_id param to the
landing page via the `@request_id` variable, which does not get
set on each request unless `sp_session[:request_id]` is blank.
If you visit the IdP via the SP, then go back to the SP and visit
the IdP again within the session validity period, `@request_id`
will be blank on the second request, and so the param will be lost.

**How**: Read the value from the session instead.