Allow user to verify phone OTP after signing out
**Why**: If a user completes the vendor verification process
but has not yet verified possession of a new phone number
via IdV, the user is prompted at next sign-in to verify their
phone. This sets up a parallel process to USPS verification,
where there is some indefinite amount of time where a profile
is in a pending state while the user verifies physical
access to their confirmed address (street or phone).

**How**: Adds a new boolean column to the `Profile` table called
`phone_confirmed` that indicates whether they have supplied
a vendor-confirmed phone number as part of their PII. There are
2 steps for all asserted PII: confirmation with vendor, and
verification via OTP. The 2nd step does not necessarily
have to happen in the same IdV session, as it cannot with
USPS delivery. Now phone supports the same scenario.