Display attributes that will be shared with SP
**Why**: After a user finishes creating their account, we want to let
them know which of their attributes will be shared with the SP.

**How**:
- Store the attributes requested by the SP in a new column in the
ServiceProviderRequests table so that they can be read on the
`sign_up/completed` page.
- Create a SamlRequestPresenter class to fetch the requested attributes
from the SAML request if present, otherwise from the SP config in the
ServiceProviders table. Then, remove any attributes we don't recognize,
and rename them to match the OpenId Connect spec. That way, we only
need to maintain one set of I18n keys.
- Extract shared code between SamlRequestPresenter and
AttributeAsserter. More of this will be done in a follow-up PR.