Return SP user to branded page when canceling
**Why**: If a user came from a Service Provider, then clicks
"Create Account", and clicks "Cancel", they should be taken back
to the branded page they saw originally, as opposed to the non-branded
home page.

Note that this only fixes the link on one page. Fixes for the links
on subsequent pages in the account creation flow will come in a separate
PR.