Improve logic in Idv::UspsMail#mail_spammed?
**Why**: If the `max_mail_events` Figaro key is not set on the server,
or if it's set to zero, and if the user has not requested any letters,
`max_events?` will return `true`, which will then cause
`updated_within_last_month?` to throw a NoMethod error because it can't
call `update_at` on a nil object.

**How**:
- Return `false` early in `mail_spammed?` if the user has not requested
any letters.
- Add `max_mail_events` and `max_mail_events_window_in_days` to the list
of required Figaro keys.
- Set `max_mail_events` to 2 in the test environment to avoid creating
more entries in the DB than is necessary.
- Use `#size` instead of `#count` for performance reasons.