Simplify OIDC authorization controller
**Why**: To avoid validating the request twice.

By validating earlier in the process (with `validate_authorize_form`),
we know that once it reaches the `index` action, it is valid. If it is
not valid, it will never reach the `index` action. This means we don't
need to validate the request both in the form's `check_submit` and
`submit` methods.

We can remove the `check_submit` method and modify the existing
`submit` method to only validate and return the FormResponse, and link
the identity in the `index` action once the user has signed in.

This is consistent with what we do with SAML: First we validate the
SAML request, and if it's valid, then we proceed to the auth action,
and once the user is fully authenticated, we link the identity.

Bonus points: This fixes the ABC complexity offense of the `index`
method.