Use Mandrill's HTTPS API to deliver emails
**Why**: For our LOA3 ATO we need to configure our firewall to only
allow HTTPS traffic. This will prevent us from sending emails over 587
with SMTPS.

Set MandrilDM API key when key is present

**Why**: Whether the Mandrill API key is present is a better indicator
of whether we want to use Mandrill than which environment we are in

Remove smtp settings from application.yml.example

**Why**: We no longer need the SMTP settings in the production now that
we are using the Mandrill API to deliver emails.