Return to branded page consistently when canceling
**Why**: Consistency is key. In this case, the cancel link on the
forgot password page was always linking to the home page.

**How**: Don't use the `shared/cancel` partial because the user will
never be signing up or going through IdV if they are accessing the
Forgot Password page, so we can directly link to the decorated session's
`cancel_link_path` which contains the logic to determine whether it
should point to the home page or the branded page.