Detect when user agent doesn't match redirect_uri
**Why**: New Relic reported errors in `CompletionsController#update`
that were due to users starting the account creation process in the
CBP Jobs iPhone app, but opened their email and continued the account
creation in a desktop browser. Because they started in the iPhone app,
the OIDC redirect URI was set to the iPhone app, so when they clicked
"Continue" on the Completions page, the redirect failed and made
another request to CompletionsController#update, but since
`session[:sp]` was deleted, the app threw an exception because it was
trying to redirect to `sp_session[:request_url]`, which was `nil`.

**How**: Compare the user agent to the redirect URI. If the URI doesn't
start with `http`, we can assume it is a mobile app. If the URI and
user agent match, i.e. if the URI is for a web app, and the user agent
is a desktop device, or if the URI is for a mobile app and the user
agent is a smartphone or tablet, then we redirect the user back to the
SP. If they don't match, we sign the user out, redirect them to the
sign in page, and display a flash message instructing them to go back
to the app on their smartphone or tablet and sign in to continue.