Require user to confirm phone with OTP during IdV
**Why**: Users should verify that they possess their phone of record
before we verify their profile. This was not happening before because we
were not actually checking that the phone number had been verified
before activating the profile.

Fix confirmations when user uses 2FA phone

**Why**: There was a bug in the ConfirmationsController. The analytics
event was checking for the precense of `phone_confirmed_at` to see if a
user had used their 2FA phone. This was a bug because
`phone_confirmed_at` is set regardless of whether the user uses their
2FA phone or a new phone. This commit fixes the controller and the
corresponding spec.

Break up idv session phone_confirmation attribute

**Why**: The use of `phone_confirmation` was confusing since both the
user and vendor must confirm the phone number. Now there is
`vendor_phone_confirmation` which is set to true when the vendor
confirms the phone, and `user_phone_confirmation` which is set to true
when the user confirms the phone.