Disable phone OTP option in select US area codes
**Why**: Twilio does not allow us to call users in some US territories
which have a +1 internation code. This commit adds frontend code that
disables the phone option and messages the user that they will have to
receive an OTP via text.

Validate that phone and otp delivery method match

**Why**: If we can't support voice for a user's phone number, and they
manage to submit the form (e.g. they have JS disabled) we want to
display an error message explaining the problem

Disable unsupported voice cofirm in phone forms

**Why**: If a user provides a phone number that is unsupported for voice
in the form to edit their 2FA phone number or the idv phone form, we
want to send them an SMS and not send them give them the option to
select phone.

Load unsupported phone area codes in data attr

**Why**: We don't want to maintain a list of unsupported area codes in
both the JS and in the ruby code.