Add OTP delivery prefence to update phone form
**Why**: So the user can select an OTP delivery preference for their new
phone number.

Merge 2FA setup and update user phone form

**Why**: With the addition of the OTP delivery preference option, both
forms are essentially identical and can be served by a single class.

Use submitted phone if phone is invalid

**Why**: The phone number formatting and validations mutates the phone
number on the user phone form. This can cause weird errors, especially
when the country code doesn't match the phone number. For example,
`+212-123-456-7890` becomes `+212-212-123-456-7890` after formatting.
This commit uses the submitted value instead of the formatted value when
the phone is invalid to prevent this.