Update Rails from 4.2.9. to 5.0.5
**Why**: To fix some security vulnerabilities in rack and rack-protection
that are only available in versions 2.0+, which depend on Rails 5+.

**How**:
- `bundle update rails sinatra` (both had to be updated at the same time
to get the latest rack)
- Specify formats in PagesController. Otherwise, visiting a nonexistent
page with a `.css` or `.png` extension would result in an exception.
- Remove the `test_after_commit` gem since its functionality is built in
to Rails 5
- Add the `rails-controller-testing` gem to be able to use expectations
such as `expect(response).to render_template(:show)`
- Update the `have_actions` custom RSpec matcher to ignore callbacks
that are tied to one of the default Rails controller actions. This is
because Rails now turns those into Procs, and we can't inspect the Proc
to see which action the callback is tied to.
- Add `.snyk` file to ignore security issues that do not affect us.

Note that there are some deprecations noted when running the specs.
These will be resolved in a separate commit that updates to Rails 5.1.