Fix deprecation warnings for Rails 5.1
**Why**: To reduce the noise when running tests, and to narrow down
the culprits when performing the upgrade to 5.1. Currently, there
are a few deprecation warnings generated by gems, but it's hard to
know which gem is causing it.

**How**:
- Use the awesome `rails5-spec-converter` gem to automatically update
all controller specs to use the new Rails 5 syntax that requires the
`params` hash.
- Remove `config.active_record.raise_in_transactional_callbacks = true`
which doesn't do anything
- Pass in a class to middleware instead of a string
- Update `config.serve_static_files` to
`config.public_file_server.enabled`
- Update `config.static_cache_control` to
`config.public_file_server.headers`
- Update `skip_before_filter` to `skip_before_action`
- Update `*_via_redirect` methods with regular `get`, `post`, etc.
and add `follow_redirect!` if necessary.
- Replace `render nothing: true` with `head`
- Use `to_h` to convert ActionController::Parameters instead of
`merge!`, which uses `to_hash` which apparently is not safe.
- Use `.where(conditions).destroy_all` instead of
`destroy_all(conditions)`
- Use `render plain` instead of `render text`