Limit OTP guesses to 3 in all contexts
**Why**: Previously, we allowed unlimited OTP guesses when confirming
a phone number because we didn't think this posed any security risks.
In hindsight this was a poor decision since it can portray our app as
being insecure and affects our reputation and confidence in the system.
We should be defaulting to safe everywhere.

**How**: Remove any conditional logic that determines whether or not
guesses should be limited. Everyone will now be limited to 3 OTP guesses
regardless of context.