Allow running any ruby 2.3 version.
Change the .ruby-version file to use `2.3` rather than `2.3.3`. This
may require users to create an alias in their ruby version config to
specify a particular point release.

For example, in rbenv you would create a symlink from `2.3` to `2.3.5`.
If you're using ruby-build and have rbenv-aliases installed, this will
happen automatically when you install new versions of ruby.

- https://github.com/tpope/rbenv-aliases

Also update the version in the Dockerfile and CircleCI config to not pin
to a specific point release.

This has no impact on login.gov servers, which don't use the
.ruby-version file. The Gemfile already allows any 2.3.* release after
2.3.3.