Specify format when rendering view in controller
**Why**: In this specific controller, we are explicitly rendering
the view (as opposed to letting the Rails magic handle the rendering)
in order to pass in local variables. By not specifying `formats: :html`,
the app will raise a `ActionView::MissingTemplate` error if someone
tries to pass in a different Accept header. To prevent unnecessary
500 errors from showing up in our logs, we can specify the HTML
format and the page will still render if any other format is passed in.